support allow_method

+++ src/methods/DELETE.cpp
@@ -3,11 +3,22 @@
 #include <errno.h>
 
 HttpResponse DELETE::execute( const HttpRequest& req ) {
-    HttpResponse resp;
+    const ServerInfo host_info = req.getHostInfo();
 
-    if (unlink(req.getPath().c_str()) < 0) {
-        // create error response
+    if (!isMethodAllowed("DELETE", req.getLocationInfo())) {
+        return HttpResponse::createErrorResponse(
+            HttpResponse::Status::METHOD_NOT_ALLOWED,
+            host_info
+        );
     }
 
-    return (resp);
+    if (unlink(req.getPath().c_str()) < 0) {
+        // TODO: support multiple errno
+        return HttpResponse::createErrorResponse(
+            HttpResponse::Status::NOT_FOUND,
+            host_info
+        );
+    }
+
+    return (HttpResponse());
 }
